move io into submodule
